+++ src/actions/downloadActions.ts
@@ -4,12 +4,17 @@ import {buildUrl, catchHandler} from './utils/fetch';
 import InvoiceModel from '../components/invoice/models/InvoiceModel';
 import {Attachment} from '../models';
 import {ClientModel} from '../components/client/models/ClientModels';
+import {getInvoiceFileName, getDownloadUrl, previewPdf, downloadAttachment} from './utils/download-helpers';
 
 
-export function getInvoiceDownloadUrl(invoice: InvoiceModel, attachment: 'pdf' | Attachment = 'pdf', downloadType?: 'preview' | 'download'): string {
+export function getInvoiceDownloadUrl(
+  invoice: InvoiceModel,
+  attachment: 'pdf' | Attachment = 'pdf',
+  downloadType?: 'preview' | 'download',
+): string {
+
   const fileType = invoice.isQuotation ? 'quotation' : 'invoice';
-  const isInvoiceAttachment = attachment === 'pdf' || attachment.type === 'pdf';
-  const fileName = isInvoiceAttachment ? getInvoiceFileName(invoice) : attachment.fileName;
+  const fileName = attachment === 'pdf' || attachment.type === 'pdf' ? getInvoiceFileName(invoice) : attachment.fileName;
   const attachmentType = attachment === 'pdf' ? 'pdf' : attachment.type;
   // return buildUrl(`/attachments/${fileType}/${invoice._id}/${attachmentType}/${encodeURIComponent(fileName)}${query}`);
 
@@ -17,63 +22,17 @@ export function getInvoiceDownloadUrl(invoice: InvoiceModel, attachment: 'pdf' |
 }
 
 
-export function getClientDownloadUrl(client: ClientModel, attachment: Attachment, downloadType: 'preview' | 'download' = 'download'): string {
+export function getClientDownloadUrl(
+  client: ClientModel,
+  attachment: Attachment,
+  downloadType: 'preview' | 'download' = 'download',
+): string {
   return getDownloadUrl('client', client._id, attachment.type, attachment.fileName, downloadType);
 }
 
 
-/** An interface you say? */
-function getDownloadUrl(
-  fileType: 'quotation' | 'invoice' | 'client',
-  _id: string,
-  attachmentType: string,
-  fileName: string, downloadType?: 'preview' | 'download',
-): string {
-
-  const query = downloadType === 'download' ? '?download=1' : '';
-  return buildUrl(`/attachments/${fileType}/${_id}/${attachmentType}/${encodeURIComponent(fileName)}${query}`);
-}
 
 
-export function invoiceReplacements(input: string, invoice: InvoiceModel): string {
-  let str = input;
-
-  const nrRegex = /\{nr:(\d+)\}/;
-  const nrMatch = str.match(nrRegex);
-  if (nrMatch) {
-    const nrSize = Math.max(parseInt(nrMatch[1], 10), invoice.number.toString().length);
-    str = str.replace(nrRegex, (`000000${invoice.number}`).slice(-nrSize));
-  }
-
-  str = str.replace(/\{nr\}/g, invoice.number.toString());
-
-  const dateRegex = /\{date:([^}]+)\}/;
-  const dateMatch = str.match(dateRegex);
-  if (dateMatch && invoice.date) {
-    const dateFormat = dateMatch[1];
-    str = str.replace(dateRegex, invoice.date.format(dateFormat));
-  }
-
-  if (str.indexOf('{orderNr}') !== -1) {
-    str = str.replace('{orderNr}', invoice.orderNr);
-  }
-
-  if (str.indexOf('{clientName}') !== -1) {
-    str = str.replace('{clientName}', invoice.client.name);
-  }
-
-  // Object.keys(data).forEach(invoiceProp => {
-  //   str = str.replace('{' + invoiceProp + '}', data[invoiceProp]);
-  // });
-
-  return str;
-}
-
-
-function getInvoiceFileName(invoice: InvoiceModel): string {
-  const {fileName} = invoice;
-  return `${invoiceReplacements(fileName, invoice)}.pdf`;
-}
 
 
 export function previewInvoice(data: InvoiceModel) {
@@ -92,25 +51,7 @@ export function previewInvoice(data: InvoiceModel) {
 
 
 
-function previewPdf(fileName: string, content: Blob): void {
-  try {
-    const blobUrl = URL.createObjectURL(content);
-    const previewWindow = window.open(blobUrl);
-    if (previewWindow) {
-      // These didn't work for setting the document.title
-      // setTimeout(() => previewWindow.document.title = fileName, 3000);
-      // previewWindow.document.write('<title>My PDF File Title</title>');
 
-      // This would work for setting a document title but need to do some additional styling
-      // previewWindow.document.write(`
-      //   <html><head><title>Your Report Title</title></head><body height="100%" width="100%">
-      //     <iframe src="' + blobUrl + '" height="100%" width="100%"></iframe>
-      //   </body></html>`);
-    }
-  } catch (err) {
-    console.error('previewPdf', err);
-  }
-}
 
 
 export function downloadInvoicesExcel(ids: string[]) {
@@ -119,7 +60,7 @@ export function downloadInvoicesExcel(ids: string[]) {
       .responseType('blob')
       .send(ids)
       .then((res) => {
-        console.log('downloaded', res);
+        console.log('downloaded', res); // eslint-disable-line
         const fileName = `invoices-${moment().format('YYYY-MM-DD')}.csv`;
         downloadAttachment(fileName, res.body);
       });
@@ -139,62 +80,3 @@ export function downloadInvoicesZip(ids: string[]) {
       });
   };
 }
-
-
-
-
-function downloadAttachment(fileName: string, content: Blob): void {
-  const link = document.createElement('a');
-  link.download = fileName;
-  const blobUrl = URL.createObjectURL(content);
-  link.href = blobUrl;
-  link.click();
-}
-
-
-
-// ATTN: This doesn't really work once deployed
-// function downloadBase64File(fileName: string, content: string): void {
-//   var link = document.createElement('a');
-//   link.download = fileName;
-//   link.target = '_blank';
-//   link.href = 'data:application/octet-stream;base64,' + content;
-//   link.click();
-// }
-
-
-
-// ATTN: this solution doesn't work on Internet Exploder
-// function openWindow(pdf: string, fileName: string): void {
-//   // GET request /attachment that just returns the bytestream
-//   // and then here:
-//   //window.open('data:application/pdf,' + escape(pdf));
-//   // (that could work right?)
-
-//   // Does work on Chrome, Firefox and Chrome
-//   var win = window.open('', '', '');
-//   if (win && win.document) {
-//     const html = `
-//       <html>
-//         <head>
-//           <title>${fileName}</title>
-//           <style>
-//             * { margin:0; padding:0 }
-//             body { margin:0; padding:0; text-align:center }
-//             #hold_my_iframe { padding:0px; margin:0 auto; width:100%; height:100% }
-//           </style>
-//         </head>
-//         <body>
-//           <table border=0 cellspacing=0 cellpadding=0 id="hold_my_iframe">
-//             <iframe src="${pdf}" width=100% height=100% marginwidth=0 marginheight=0 frameborder=0></iframe>
-//           </table>
-//         </body>
-//       </html>`;
-
-//     win.document.write(html);
-//     win.document.title = fileName;
-
-//   } else {
-//     failure(t('controls.popupBlocker'), t('controls.popupBlockerTitle'), 8000);
-//   }
-// }
