+++ src/components/controls/attachments/AddAttachmentPopup.tsx
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {connect} from 'react-redux';
 import {FormLabel, FormGroup, Alert} from 'react-bootstrap';
 import {t} from '../../utils';
-import {Popup, SimpleCreatableSelect, PopupButton} from '../../controls';
+import {Popup, SimpleCreatableSelect, PopupButton} from '../../controls'; // TODO: Dependency cycle detected
 import {Attachment} from '../../../models';
 import {ConfacState} from '../../../reducers/app-state';
 import {AttachmentDropzone} from './AttachmentDropzone';
