+++ src/components/controls/attachments/AttachmentsForm.tsx
@@ -1,18 +1,10 @@
 import React, {Component} from 'react';
 import {connect} from 'react-redux';
-import {Row, Col} from 'react-bootstrap';
 import {Attachment, IAttachment} from '../../../models';
-import {HeaderWithEditIcon} from '../Headers';
-import {t} from '../../utils';
-import {AddIcon, ConfirmedDeleteIcon} from '../Icon';
-import {AddAttachmentPopup} from './AddAttachmentPopup';
 import {updateAttachment, deleteAttachment} from '../../../actions/index';
-import {ProposedAttachmentsDropzones} from './ProposedAttachmentsDropzones';
-import {AttachmentForm} from './AttachmentForm';
-
+import {AbstractAttachmentsForm} from './AbstractAttachmentsForm';
 
 import './attachments.scss';
-import {AttachmentPreviewIcon} from '../../controls';
 
 
 export type AttachmentsFormProps = {
@@ -22,6 +14,7 @@ export type AttachmentsFormProps = {
 }
 
 
+// eslint-disable-next-line react/prefer-stateless-function
 export class AttachmentsFormComponent extends Component<AttachmentsFormProps> {
   render() {
     const {model} = this.props;
@@ -30,12 +23,13 @@ export class AttachmentsFormComponent extends Component<AttachmentsFormProps> {
     }
 
 
-    const modelType = model.getType ? model.getType() : 'client';
+    // eslint-disable-next-line dot-notation
+    const modelType = model['getType'] ? model['getType']() : 'client';
     return (
       <AbstractAttachmentsForm
         attachments={model.attachments}
         onDelete={(att: Attachment) => this.props.deleteAttachment(model, modelType, att)}
-        onAdd={(att) => this.props.updateAttachment(model, modelType, att)}
+        onAdd={att => this.props.updateAttachment(model, modelType, att)}
         model={model}
         modelType={modelType}
       />
@@ -47,7 +41,7 @@ export const AttachmentsForm = connect(null, {updateAttachment, deleteAttachment
 
 
 
-type AbstractAttachmentsFormProps = {
+export type AbstractAttachmentsFormProps = {
   attachments: Attachment[],
   onDelete: Function,
   onAdd: ({file: File, type: string}) => void,
@@ -55,84 +49,11 @@ type AbstractAttachmentsFormProps = {
   modelType: 'invoice' | 'client',
 }
 
-type AbstractAttachmentsFormState = {
+export type AbstractAttachmentsFormState = {
   /** Show the attachment upload popup */
   isOpen: boolean,
   /** Show delete icons on the attachments */
   isFormOpen: boolean,
   /** The attachment.type to show the preview icon */
   hoverId: string | null,
-}
-
-export class AbstractAttachmentsForm extends Component<AbstractAttachmentsFormProps, AbstractAttachmentsFormState> {
-  constructor(props: AbstractAttachmentsFormProps) {
-    super(props);
-    this.state = {
-      isOpen: false,
-      isFormOpen: false,
-      hoverId: null,
-    };
-  }
-
-  render() {
-    const attachments = this.props.attachments.filter((att) => att.type !== 'pdf');
-    const canDeleteAttachments = attachments.length > 0;
-
-    return (
-      <Row className="tst-attachments attachments-form">
-        <Col sm={12}>
-          <HeaderWithEditIcon
-            size={2}
-            label={t('invoice.attachments')}
-            editIconVisible={canDeleteAttachments}
-            onEditClick={() => this.setState({isFormOpen: !this.state.isFormOpen})}
-          />
-
-          <AddIcon
-            style={{marginTop: 0, marginLeft: 16, marginBottom: 26}}
-            onClick={() => this.setState({isOpen: true})}
-            label={t('invoice.attachmentsAdd')}
-            size={1}
-            data-tst="add-attachment"
-          />
-
-          <AddAttachmentPopup
-            isOpen={this.state.isOpen}
-            attachments={this.props.attachments}
-            onClose={() => this.setState({isOpen: false})}
-            onAdd={(att: { file: File, type: string }) => this.props.onAdd(att)}
-          />
-        </Col>
-
-        {attachments.map((att) => (
-          <Col lg={4} md={6} key={att.type} onMouseEnter={() => this.setState({hoverId: att.type})} onMouseLeave={() => this.setState({hoverId: null})}>
-            <AttachmentForm
-              model={this.props.model}
-              modelType={this.props.modelType}
-              attachment={att}
-            >
-              {this.state.isFormOpen && (
-                <div className="delete">
-                  <ConfirmedDeleteIcon
-                    title={t('attachment.deleteTitle')}
-                    onClick={this.props.onDelete.bind(this, att)}
-                    data-tst={`att-delete-${att.type}`}
-                  >
-                    {t('attachment.deletePopup')}
-                  </ConfirmedDeleteIcon>
-                </div>
-              )}
-              {!this.state.isFormOpen && this.state.hoverId === att.type && (
-                <div className="delete">
-                  <AttachmentPreviewIcon model={this.props.model} modelType={this.props.modelType} attachment={att} />
-                </div>
-              )}
-            </AttachmentForm>
-          </Col>
-        ))}
-
-        <ProposedAttachmentsDropzones model={this.props.model} modelType={this.props.modelType} />
-      </Row>
-    );
-  }
 }
