+++ src/reducers/consultant-reducers.ts
@@ -5,9 +5,10 @@ export const consultants = (state: ConsultantModel[] = [], action): ConsultantMo
   switch (action.type) {
     case ACTION_TYPES.CONSULTANTS_FETCHED:
       return action.consultants;
-    case ACTION_TYPES.CONSULTANT_UPDATE:
+    case ACTION_TYPES.CONSULTANT_UPDATE: {
       const newState: ConsultantModel[] = [...state, action.consultant];
       return newState;
+    }
     default:
       return state;
   }
