+++ src/reducers/project-reducers.ts
@@ -5,9 +5,10 @@ export const projects = (state: ProjectModel[] = [], action): ProjectModel[] => 
   switch (action.type) {
     case ACTION_TYPES.PROJECTS_FETCHED:
       return action.projects;
-    case ACTION_TYPES.PROJECT_UPDATE:
+    case ACTION_TYPES.PROJECT_UPDATE: {
       const newState: ProjectModel[] = [...state, action.project];
       return newState;
+    }
     default:
       return state;
   }
